+++ velcro-core/vcore/io/path/path_parser.inl
@@ -649,8 +649,8 @@ namespace V::IO::parser
     /// Path is using FNV-1a algorithm 64 bit version.
     inline size_t HashSegment(VStd::string_view pathSegment, bool hashExactPath)
     {
-        size_t hash = 14695981039346656037ULL;
-        constexpr size_t fnvPrime = 1099511628211ULL;
+        size_t hash = (size_t)14695981039346656037ULL;
+        constexpr size_t fnvPrime = (size_t)1099511628211ULL;
 
         for (const char first : pathSegment)
         {
